Tested NWS API calls, browsed data
Calls work, although ./observations/latest for current weather provides
many stations in a grid, and some of them aren't returning values. This
Makes it hard to know which station you should pick, and how to tell if
it is active. Maybe get all the data, and iterate to find the first
usable value, or an average of them all?